refactor(mt#797): split ai.ts and config.ts into focused sub-modules
ai.ts (763 lines) → ai/completion-commands.ts, ai/model-cache-commands.ts,
ai/provider-commands.ts, ai/shared-helpers.ts; barrel is now ~25 lines.

config.ts (762 lines) → config/helpers.ts, config/list-show-commands.ts,
config/get-set-commands.ts, config/validate-doctor-commands.ts; barrel is now ~25 lines.

All original public exports preserved (registerAiCommands, registerConfigCommands).
